requirements.txt
alembic==1.7.4
attrs==21.2.0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
decorator==5.1.0
flasgger==0.9.5
Flask==2.0.2
Flask-JWT-Extended==4.3.1
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-core==2.2.0
google-cloud-datastore==2.4.0
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
grpcio==1.41.1
grpcio-status==1.41.1
gunicorn==20.1.0
idna==3.3
install==1.3.4
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==4.1.2
libcst==0.3.21
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.14.0
marshmallow-sqlalchemy==0.26.1
mistune==0.8.4
mypy-extensions==0.4.3
proto-plus==1.19.8
protobuf==3.19.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pyrsistent==0.18.0
python-dotenv==0.19.1
PyYAML==6.0
requests==2.26.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.26
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
validators==0.18.2
Werkzeug==2.0.2
